refactor: auto-split test_putting_simulation.py
