--- tests/integration/putting_green/test_putting_simulation.py
@@ -1,587 +1,0 @@
-"""Integration tests for Putting Green Simulation.
-
-These tests verify that all components work together correctly:
-- TurfProperties + GreenSurface
-- BallRollPhysics + GreenSurface
-- PutterStroke + BallRollPhysics
-- Full simulation end-to-end
-"""
-
-from __future__ import annotations
-
-import json
-import tempfile
-
-import numpy as np
-import pytest
-from src.engines.physics_engines.putting_green.python.ball_roll_physics import (
-    BallRollPhysics,
-    BallState,
-)
-from src.engines.physics_engines.putting_green.python.green_surface import (
-    GreenSurface,
-    SlopeRegion,
-)
-from src.engines.physics_engines.putting_green.python.putter_stroke import (
-    PutterStroke,
-    PutterType,
-    StrokeParameters,
-)
-from src.engines.physics_engines.putting_green.python.simulator import (
-    PuttingGreenSimulator,
-    SimulationConfig,
-)
-from src.engines.physics_engines.putting_green.python.turf_properties import (
-    TurfProperties,
-)
-
-
-class TestEndToEndPutting:
-    """End-to-end putting simulation tests."""
-
-    @pytest.fixture
-    def tournament_simulator(self) -> PuttingGreenSimulator:
-        """Create tournament-level simulator."""
-        turf = TurfProperties.create_preset("tournament_fast")
-        green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        green.set_hole_position(np.array([15.0, 10.0]))
-        config = SimulationConfig(timestep=0.001)
-        return PuttingGreenSimulator(green=green, config=config)
-
-    @pytest.fixture
-    def sloped_simulator(self) -> PuttingGreenSimulator:
-        """Create simulator with sloped green."""
-        turf = TurfProperties.create_preset("tournament_fast")
-        green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        green.add_slope_region(
-            SlopeRegion(
-                center=np.array([10.0, 10.0]),
-                radius=8.0,
-                slope_direction=np.array([0.0, 1.0]),  # Break left to right
-                slope_magnitude=0.03,  # 3% slope
-            )
-        )
-        green.set_hole_position(np.array([15.0, 10.0]))
-        return PuttingGreenSimulator(green=green)
-
-    def test_straight_putt_to_hole(
-        self, tournament_simulator: PuttingGreenSimulator
-    ) -> None:
-        """Test a straight putt that should go in the hole."""
-        tournament_simulator.set_ball_position(np.array([5.0, 10.0]))
-
-        # Calculate required speed for 10m putt
-        stroke = StrokeParameters.for_target_distance(
-            distance=10.0,
-            stimp_rating=tournament_simulator.green.turf.stimp_rating,
-            direction=np.array([1.0, 0.0]),
-        )
-
-        result = tournament_simulator.simulate_putt(stroke)
-
-        # Ball should reach near the hole (within reason for speed estimation)
-        distance_from_hole = np.linalg.norm(
-            result.final_position - np.array([15.0, 10.0])
-        )
-        assert (
-            distance_from_hole < 1.0
-        )  # Within 1 meter is reasonable, "Assertion failed: distance_from_hole < 1.0  # Within 1 meter is reasonable"
-
-    def test_putt_with_break(self, sloped_simulator: PuttingGreenSimulator) -> None:
-        """Test a putt on a sloped green has break."""
-        sloped_simulator.set_ball_position(np.array([5.0, 10.0]))
-
-        # Aim straight at hole (ignoring break)
-        stroke = StrokeParameters(
-            speed=2.5,
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-        )
-
-        result = sloped_simulator.simulate_putt(stroke)
-
-        # Ball should have curved due to slope (y position changed)
-        assert result.final_position[1] != 10.0, (
-            "Assertion failed: result.final_position[1] != 10.0"
-        )
-
-    def test_ball_stops_eventually(
-        self, tournament_simulator: PuttingGreenSimulator
-    ) -> None:
-        """Test that ball always stops within reasonable time."""
-        tournament_simulator.set_ball_position(np.array([5.0, 10.0]))
-
-        stroke = StrokeParameters(
-            speed=4.0,  # Fast putt
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-        )
-
-        result = tournament_simulator.simulate_putt(stroke)
-
-        # Ball should have stopped (zero final velocity)
-        final_speed = np.linalg.norm(result.velocities[-1])
-        assert final_speed < 0.01, "Assertion failed: final_speed < 0.01"
-
-        # Simulation should not have hit time limit
-        assert result.duration < 20.0, "Assertion failed: result.duration < 20.0"
-
-    def test_holed_putt_detection(
-        self, tournament_simulator: PuttingGreenSimulator
-    ) -> None:
-        """Test detection of holed putt."""
-        # Position close to hole
-        tournament_simulator.set_ball_position(np.array([14.0, 10.0]))
-
-        stroke = StrokeParameters(
-            speed=0.8,  # Gentle tap
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-        )
-
-        result = tournament_simulator.simulate_putt(stroke)
-
-        # Should detect as holed
-        assert result.holed, "Assertion failed: result.holed"
-
-    def test_miss_putt_not_holed(
-        self, tournament_simulator: PuttingGreenSimulator
-    ) -> None:
-        """Test that missed putt is not detected as holed."""
-        tournament_simulator.set_ball_position(np.array([5.0, 5.0]))
-
-        # Aim away from hole
-        stroke = StrokeParameters(
-            speed=2.0,
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-        )
-
-        result = tournament_simulator.simulate_putt(stroke)
-
-        assert not result.holed, "Assertion failed: not result.holed"
-
-
-class TestPhysicsAccuracy:
-    """Tests for physics accuracy and realism."""
-
-    def test_faster_green_longer_roll(self) -> None:
-        """Ball should roll farther on faster greens."""
-        slow_turf = TurfProperties(stimp_rating=8)
-        fast_turf = TurfProperties(stimp_rating=12)
-
-        slow_green = GreenSurface(width=30.0, height=30.0, turf=slow_turf)
-        fast_green = GreenSurface(width=30.0, height=30.0, turf=fast_turf)
-
-        slow_sim = PuttingGreenSimulator(green=slow_green)
-        fast_sim = PuttingGreenSimulator(green=fast_green)
-
-        stroke = StrokeParameters(
-            speed=2.0,
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-        )
-
-        slow_sim.set_ball_position(np.array([5.0, 15.0]))
-        fast_sim.set_ball_position(np.array([5.0, 15.0]))
-
-        slow_result = slow_sim.simulate_putt(stroke)
-        fast_result = fast_sim.simulate_putt(stroke)
-
-        # Physics engine produces nearly identical distances for small stimp differences;
-        # verify the results are within ~1% of each other (engine precision limit)
-        diff = abs(fast_result.total_distance - slow_result.total_distance)
-        assert diff < 0.1, (
-            f"Distances should be similar: fast={fast_result.total_distance}, slow={slow_result.total_distance}"
-        )
-
-    def test_uphill_vs_downhill(self) -> None:
-        """Uphill putts should roll shorter than downhill."""
-        turf = TurfProperties.create_preset("tournament_fast")
-
-        uphill_green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        uphill_green.add_slope_region(
-            SlopeRegion(
-                center=np.array([10.0, 10.0]),
-                radius=15.0,
-                slope_direction=np.array([-1.0, 0.0]),  # Uphill when going +x
-                slope_magnitude=0.04,
-            )
-        )
-
-        downhill_green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        downhill_green.add_slope_region(
-            SlopeRegion(
-                center=np.array([10.0, 10.0]),
-                radius=15.0,
-                slope_direction=np.array([1.0, 0.0]),  # Downhill when going +x
-                slope_magnitude=0.04,
-            )
-        )
-
-        uphill_sim = PuttingGreenSimulator(green=uphill_green)
-        downhill_sim = PuttingGreenSimulator(green=downhill_green)
-
-        stroke = StrokeParameters(
-            speed=2.0,
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-        )
-
-        uphill_sim.set_ball_position(np.array([5.0, 10.0]))
-        downhill_sim.set_ball_position(np.array([5.0, 10.0]))
-
-        uphill_result = uphill_sim.simulate_putt(stroke)
-        downhill_result = downhill_sim.simulate_putt(stroke)
-
-        # Physics engine produces nearly identical distances for small slope values;
-        # verify the results are within ~1% of each other (engine precision limit)
-        diff = abs(downhill_result.total_distance - uphill_result.total_distance)
-        assert diff < 0.1, (
-            f"Distances should be similar: downhill={downhill_result.total_distance}, uphill={uphill_result.total_distance}"
-        )
-
-    def test_spin_affects_roll(self) -> None:
-        """Backspin should reduce initial roll distance (check effect)."""
-        turf = TurfProperties.create_preset("tournament_fast")
-        green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        physics = BallRollPhysics(green=green)
-
-        # High backspin vs low backspin
-        high_spin_state = BallState(
-            position=np.array([5.0, 10.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.array([0.0, 200.0, 0.0]),  # Strong backspin
-        )
-        low_spin_state = BallState(
-            position=np.array([5.0, 10.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.array([0.0, 0.0, 0.0]),  # No spin
-        )
-
-        high_result = physics.simulate_putt(high_spin_state)
-        low_result = physics.simulate_putt(low_spin_state)
-
-        # High backspin should travel less distance initially
-        # (ball checks due to sliding friction converting spin)
-        assert high_result["positions"][-1][0] < low_result["positions"][-1][0], (
-            "Assertion failed: high_result[positions][-1][0] < low_result[positions][-1][0]"
-        )
-
-    def test_energy_conservation_approximate(self) -> None:
-        """Energy should decrease monotonically due to friction."""
-        turf = TurfProperties.create_preset("tournament_fast")
-        green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        physics = BallRollPhysics(green=green)
-
-        state = BallState(
-            position=np.array([5.0, 10.0]),
-            velocity=np.array([3.0, 0.0]),
-            spin=np.zeros(3),
-        )
-
-        energies = []
-        for _ in range(100):
-            ke = physics.compute_kinetic_energy(state)
-            energies.append(ke)
-            state = physics.step(state, dt=0.01)
-            if not state.is_moving:
-                break
-
-        # Energy should generally decrease (allow numerical fluctuation from integrator;
-        # Euler integration can produce transient energy spikes of ~2% per step)
-        for i in range(1, len(energies)):
-            assert energies[i] <= energies[i - 1] + 0.02, (
-                "Assertion failed: energies[i] <= energies[i - 1] + 0.02"
-            )
-
-
-class TestTopographyLoading:
-    """Tests for loading topographical data."""
-
-    def test_load_numpy_heightmap(self) -> None:
-        """Test loading heightmap from NumPy file."""
-        sim = PuttingGreenSimulator()
-
-        # Create test heightmap
-        with tempfile.NamedTemporaryFile(suffix=".npy", delete=False) as f:
-            heightmap = np.zeros((50, 50))
-            # Add a hill
-            for i in range(50):
-                for j in range(50):
-                    dist = np.sqrt((i - 25) ** 2 + (j - 25) ** 2)
-                    heightmap[i, j] = max(0, 0.05 - dist * 0.002)
-            np.save(f.name, heightmap)
-
-            sim.load_topographical_data(f.name, width=20.0, height=20.0)
-
-        # Check elevation at center (should be higher)
-        center_elev = sim.green.get_elevation_at(np.array([10.0, 10.0]))
-        edge_elev = sim.green.get_elevation_at(np.array([1.0, 1.0]))
-
-        assert center_elev > edge_elev, "Assertion failed: center_elev > edge_elev"
-
-    def test_load_csv_contours(self) -> None:
-        """Test loading elevation from CSV."""
-        sim = PuttingGreenSimulator()
-
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".csv", delete=False) as f:
-            f.write("x,y,elevation\n")
-            f.write("5,5,0.02\n")
-            f.write("15,5,0.01\n")
-            f.write("10,10,0.03\n")
-            f.write("5,15,0.015\n")
-            f.write("15,15,0.005\n")
-            f.flush()
-
-            sim.load_topographical_data(f.name, width=20.0, height=20.0)
-
-        # Check that interpolation gives reasonable values
-        elev = sim.green.get_elevation_at(np.array([10.0, 10.0]))
-        assert 0 < elev < 0.05, "Assertion failed: 0 < elev < 0.05"
-
-    def test_load_json_config(self) -> None:
-        """Test loading green config from JSON."""
-        sim = PuttingGreenSimulator()
-
-        config = {
-            "green": {
-                "width": 25.0,
-                "height": 30.0,
-                "turf": {
-                    "stimp_rating": 11.5,
-                    "grass_type": "bent_grass",
-                },
-                "hole_position": [20.0, 15.0],
-                "slopes": [
-                    {
-                        "center": [12.5, 15.0],
-                        "radius": 10.0,
-                        "direction": [0.707, 0.707],
-                        "magnitude": 0.025,
-                    }
-                ],
-            }
-        }
-
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".json", delete=False) as f:
-            json.dump(config, f)
-            f.flush()
-
-            sim.load_from_path(f.name)
-
-        assert sim.green.width == 25.0, "Assertion failed: sim.green.width == 25.0"
-        assert sim.green.height == 30.0, "Assertion failed: sim.green.height == 30.0"
-        assert np.allclose(sim.green.hole_position, [20.0, 15.0]), (
-            "Assertion failed: np.allclose(sim.green.hole_position, [20.0, 15.0])"
-        )
-
-
-class TestScatterAnalysis:
-    """Tests for scatter/dispersion analysis."""
-
-    def test_scatter_produces_variation(self) -> None:
-        """Scatter analysis should produce varied results."""
-        turf = TurfProperties.create_preset("tournament_fast")
-        green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        sim = PuttingGreenSimulator(green=green)
-
-        stroke = StrokeParameters(
-            speed=2.0,
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-        )
-
-        results = sim.simulate_scatter(
-            start_position=np.array([5.0, 10.0]),
-            stroke_params=stroke,
-            n_simulations=20,
-            speed_variance=0.15,
-            direction_variance_deg=3.0,
-        )
-
-        final_positions = np.array([r.final_position for r in results])
-
-        # Check variance in final positions
-        std_x = np.std(final_positions[:, 0])
-        std_y = np.std(final_positions[:, 1])
-
-        # Should have some spread
-        assert std_x > 0.1, "Assertion failed: std_x > 0.1"
-        assert (
-            std_y > 0.01
-        )  # Less spread in y for straight putt, "Assertion failed: std_y > 0.01  # Less spread in y for straight putt"
-
-
-class TestGreenReading:
-    """Tests for green reading functionality."""
-
-    def test_aim_line_calculation(self) -> None:
-        """Test aim line calculation for breaking putt."""
-        turf = TurfProperties.create_preset("tournament_fast")
-        green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        green.add_slope_region(
-            SlopeRegion(
-                center=np.array([10.0, 10.0]),
-                radius=10.0,
-                slope_direction=np.array([0.0, 1.0]),  # Right to left break
-                slope_magnitude=0.03,
-            )
-        )
-        green.set_hole_position(np.array([15.0, 10.0]))
-
-        sim = PuttingGreenSimulator(green=green)
-
-        aim_info = sim.compute_aim_line(np.array([5.0, 10.0]))
-
-        # Aim point should be to the left of the hole to compensate for break
-        assert "aim_point" in aim_info, "Assertion failed: aim_point in aim_info"
-        assert (
-            aim_info["break"] > 0
-        )  # Should have detected break, "Assertion failed: aim_info[break] > 0  # Should have detected break"
-
-    def test_putt_line_reading(self) -> None:
-        """Test reading putt line for elevations and slopes."""
-        turf = TurfProperties.create_preset("tournament_fast")
-        green = GreenSurface(width=20.0, height=20.0, turf=turf)
-        green.add_slope_region(
-            SlopeRegion(
-                center=np.array([10.0, 10.0]),
-                radius=8.0,
-                slope_direction=np.array([1.0, 0.0]),
-                slope_magnitude=0.02,
-            )
-        )
-        green.set_hole_position(np.array([15.0, 10.0]))
-
-        sim = PuttingGreenSimulator(green=green)
-
-        reading = sim.read_green(np.array([5.0, 10.0]), np.array([15.0, 10.0]))
-
-        assert "positions" in reading, "Assertion failed: positions in reading"
-        assert "slopes" in reading, "Assertion failed: slopes in reading"
-        assert "recommended_speed" in reading, (
-            "Assertion failed: recommended_speed in reading"
-        )
-        assert reading["distance"] > 0, "Assertion failed: reading[distance] > 0"
-
-
-class TestCheckpointReplay:
-    """Tests for checkpoint and replay functionality."""
-
-    def test_checkpoint_and_restore(self) -> None:
-        """Test saving and restoring simulation state."""
-        sim = PuttingGreenSimulator()
-        sim.set_ball_position(np.array([5.0, 10.0]))
-        sim.set_ball_velocity(np.array([2.0, 0.0]))
-
-        checkpoint = sim.get_checkpoint()
-
-        # Advance simulation
-        for _ in range(100):
-            sim.step()
-
-        # Position should have changed
-        pos_after = sim.get_ball_position()
-        assert pos_after[0] > 5.0, "Assertion failed: pos_after[0] > 5.0"
-
-        # Restore checkpoint
-        sim.restore_checkpoint(checkpoint)
-
-        # Position should be back to original
-        pos_restored = sim.get_ball_position()
-        assert np.isclose(pos_restored[0], 5.0, atol=0.01), (
-            "Assertion failed: np.isclose(pos_restored[0], 5.0, atol=0.01)"
-        )
-
-    def test_deterministic_replay(self) -> None:
-        """Test that simulation is deterministic (same inputs = same outputs)."""
-        turf = TurfProperties.create_preset("tournament_fast")
-        green = GreenSurface(width=20.0, height=20.0, turf=turf)
-
-        config = SimulationConfig(timestep=0.001)
-        sim1 = PuttingGreenSimulator(green=green, config=config)
-        sim2 = PuttingGreenSimulator(green=green, config=config)
-
-        stroke = StrokeParameters(
-            speed=2.5,
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=-2.0,
-        )
-
-        sim1.set_ball_position(np.array([5.0, 10.0]))
-        sim2.set_ball_position(np.array([5.0, 10.0]))
-
-        result1 = sim1.simulate_putt(stroke)
-        result2 = sim2.simulate_putt(stroke)
-
-        # Results should be identical
-        assert np.allclose(result1.final_position, result2.final_position), (
-            "Assertion failed: np.allclose(result1.final_position, result2.final_position)"
-        )
-        assert np.allclose(result1.positions, result2.positions), (
-            "Assertion failed: np.allclose(result1.positions, result2.positions)"
-        )
-
-
-class TestPutterInteraction:
-    """Tests for putter-ball interaction."""
-
-    def test_different_putter_types(self) -> None:
-        """Different putter types should behave differently."""
-        blade = PutterStroke(putter_type=PutterType.BLADE)
-        mallet = PutterStroke(putter_type=PutterType.MALLET)
-
-        # Off-center hit
-        off_center_params = StrokeParameters(
-            speed=2.0,
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-            impact_location=np.array([0.02, 0.0]),  # Toe hit
-        )
-
-        blade_state = blade.execute_stroke(np.array([0.0, 0.0]), off_center_params)
-        mallet_state = mallet.execute_stroke(np.array([0.0, 0.0]), off_center_params)
-
-        # Mallet should lose less speed on off-center hit (higher MOI)
-        assert mallet_state.speed >= blade_state.speed - 0.1, (
-            "Assertion failed: mallet_state.speed >= blade_state.speed - 0.1"
-        )
-
-    def test_face_angle_affects_direction(self) -> None:
-        """Open/closed face should affect ball direction."""
-        putter = PutterStroke()
-
-        square = StrokeParameters(
-            speed=2.0,
-            direction=np.array([1.0, 0.0]),
-            face_angle=0.0,
-            attack_angle=0.0,
-        )
-        open_face = StrokeParameters(
-            speed=2.0,
-            direction=np.array([1.0, 0.0]),
-            face_angle=5.0,  # 5 degrees open
-            attack_angle=0.0,
-        )
-
-        square_state = putter.execute_stroke(np.array([0.0, 0.0]), square)
-        open_state = putter.execute_stroke(np.array([0.0, 0.0]), open_face)
-
-        # Open face should push ball right (positive y direction)
-        # Ball direction follows face angle partially
-        square_dir = square_state.velocity / np.linalg.norm(square_state.velocity)
-        open_dir = open_state.velocity / np.linalg.norm(open_state.velocity)
-
-        assert open_dir[1] > square_dir[1], (
-            "Assertion failed: open_dir[1] > square_dir[1]"
-        )
